+++ examples/debit_card_pull/debit_pull_test.go
@@ -17,6 +17,9 @@ func TestDebitPullWithRefund(t *testing.T) {
 	// However, it is recommended to load the credentials from the
 	// configuration file.
 
+	// The account facilitating the transfer
+	partnerAccountID := "5352b013-ae58-4a63-8a3f-97f316a917cf" // example
+
 	mc, err := moov.NewClient(moov.WithCredentials(moov.CredentialsFromEnv()))
 	require.NoError(t, err)
 
@@ -78,6 +81,7 @@ func TestDebitPullWithRefund(t *testing.T) {
 	// Step 6: create transfer
 	completedTransfer, _, err := mc.CreateTransfer(
 		ctx,
+		partnerAccountID,
 		moov.CreateTransfer{
 			Source: moov.CreateTransfer_Source{
 				PaymentMethodID: sourcePaymentMethodID,
@@ -108,6 +112,7 @@ func TestDebitPullWithRefund(t *testing.T) {
 	// Step 7: refund transfer
 	refund, _, err := mc.RefundTransfer(
 		ctx,
+		partnerAccountID,
 		completedTransfer.TransferID,
 		moov.CreateRefund{
 			Amount: 97,