+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## v0.10.0 (Released 2025-03-04)
+
+BREAKING CHANGES
+
+- transfers: update methods to use the `/accounts/{accountID}`-prefixed endpoints
+- cards: update `CardDetails` model to use new `CardTransactionStatus` enum
+
+ADDITIONS
+
+- transfers: add cancellation API endpoints and client methods
+
 ## v0.9.0 (Released 2025-01-10)
 
 ADDITIONS